Vastly improve performance and add support for ObjC Kotlin classes in Swift.

+++ LLDBPlugin/run.py
@@ -3,43 +3,54 @@
 import os
 import subprocess
 
-from lldb import SBDebugger, SBTarget, LLDB_ARCH_DEFAULT, SBEvent, SBListener, SBProcess, eStateStopped
+import lldb
 
-gradle_invoke = subprocess.Popen(['./gradlew', 'linkDebugExecutableMacosArm64'], cwd='test_project')
-gradle_invoke.wait()
+gradle_invoke = subprocess.Popen(['./gradlew', 'compileSwift'], cwd='test_project')
+gradle_exit_code = gradle_invoke.wait()
 
-exe = 'test_project/build/bin/macosArm64/debugExecutable/test_project.kexe'
+if gradle_exit_code != 0:
+    exit(gradle_exit_code)
+
+exe = 'test_project/build/swift/app'
 
 # Initialize the debugger before making any API calls.
-SBDebugger.Initialize()
+lldb.SBDebugger.Initialize()
 # Create a new debugger instance in your module if your module
 # can be run from the command line. When we run a script from
 # the command line, we won't have any debugger object in
 # lldb.debugger, so we can just create it if it will be needed
-debugger: SBDebugger = SBDebugger.Create()
+debugger: lldb.SBDebugger = lldb.SBDebugger.Create()
 
 # When we step or continue, don't return from the function until the process
 # stops. Otherwise we would have to handle the process events ourselves which, while doable is
 # a little tricky.  We do this by setting the async mode to false.
 debugger.SetAsync(False)
 
+# debugger.HandleCommand('log enable lldb default')
+
 # Import our module
 debugger.HandleCommand('command script import touchlab_kotlin_lldb')
 
-target: SBTarget = debugger.CreateTargetWithFileAndArch(exe, LLDB_ARCH_DEFAULT)
+target: lldb.SBTarget = debugger.CreateTargetWithFileAndArch(exe, lldb.LLDB_ARCH_DEFAULT)
 
 if target:
-    target.BreakpointCreateByLocation('main.kt', 13)
+    # target.BreakpointCreateByLocation("main.swift", 14)
+    target.BreakpointCreateByLocation('main.kt', 47)
 
-    process = target.LaunchSimple(None, None, os.getcwd())
+    process: lldb.SBProcess = target.LaunchSimple(None, None, os.getcwd())
 
     if process:
+        import time
+        start = time.perf_counter()
         debugger.HandleCommand('fr v --ptr-depth 16')
+        print('HandleCommand took {:.6}s'.format(time.perf_counter() - start))
+        process.Continue()
         # debugger.HandleCommand('fr v --ptr-depth 16 -- data')
+        debugger.HandleCommand('fr v --ptr-depth 16')
 
         process.Kill()
 
 # Finally, dispose of the debugger you just made.
-SBDebugger.Destroy(debugger)
+lldb.SBDebugger.Destroy(debugger)
 # Terminate the debug session
-SBDebugger.Terminate()
+lldb.SBDebugger.Terminate()
